\name{Commodities}
\alias{Commodities}
\docType{data}
\title{ Commodity prices }
\description{
Commodity prices on successive business days, Chicago Exchange
These data exhibit classic randow walk behavior.
}
\usage{data(Commodities)}
\format{
  The format is:
List of 5
 $ gold:'data.frame':	97 obs. of  3 variables:
  ..$ close: num [1:97] 700 671 680 677 690 ...
  ..$ high : num [1:97] 714 698 683 682 692 ...
  ..$ low  : num [1:97] 700 669 664 676 684 ...
 $ feed:'data.frame':	95 obs. of  3 variables:
  ..$ close: num [1:95] 79 79 78.6 79.9 79.3 ...
  ..$ high : num [1:95] 80 79.5 79.2 79.9 79.8 ...
  ..$ low  : num [1:95] 79 78.5 78.6 78.8 79.3 ...
 $ port:'data.frame':	99 obs. of  3 variables:
  ..$ close: num [1:99] 57.7 56.8 57.5 57 59 ...
  ..$ high : num [1:99] 59.9 57.5 58 58.1 59 ...
  ..$ low  : num [1:99] 57.2 56.4 55.1 56.8 56.4 ...
 $ soy :'data.frame':	99 obs. of  3 variables:
  ..$ close: num [1:99] 766 790 804 794 824 ...
  ..$ high : num [1:99] 788 791 805 808 824 ...
  ..$ low  : num [1:99] 764 764 778 792 809 ...
 $ us  :'data.frame':	100 obs. of  3 variables:
  ..$ close: num [1:100] 91.6 91.6 91.4 91.4 91.2 ...
  ..$ high : num [1:100] 91.9 91.7 91.6 91.4 91.5 ...
  ..$ low  : num [1:100] 91.6 91.5 91.3 91.3 91.1 ...
}
\details{
Data from 1981. feed: April; gold: June, pork: March, us: March
}
\source{
I obtained these data from a broker.
}

\examples{
dim(Commodities$gold)
dimnames(Commodities$gold)[[2]]
TimeSeriesPlot(Commodities$gold$close)
}
\keyword{datasets}
