\name{FromSymmetricStorageUpper}
\alias{FromSymmetricStorageUpper}
\title{ Converts a Matrix from Symmetric Storage Mode to Regular Format }
\description{
 Utility function.
}
\usage{
FromSymmetricStorageUpper(x)
}
\arguments{
  \item{x}{ a vector which represents a matrix in upper triangular form }
}
\value{
symmetric matrix
}
\author{ A.I. McLeod }


\examples{
FromSymmetricStorageUpper(1:5)
}
\keyword{ ts }
\keyword{ array }
