\name{GetB}
\alias{GetB}
\title{ Internal Utility Function  }
\description{
 The user would not normally use this function. 
 The function is needed for exact mle for mean.
 Used in Get1G which is called from GetARMeanMLE.
}
\usage{
GetB(phi)
}
\arguments{
  \item{phi}{ vector of AR coefficients }
}
\keyword{ ts }

