\name{SeriesB}
\alias{SeriesB}
\docType{data}
\title{ Series B}

\description{
closing price of common stock, daily, May 17 1961 to November 2 1962
}

\usage{data(SeriesB)}

\format{
  The format is:
 num [1:369] 460 457 452 459 462 459 463 479 493 490 ...
}

\source{
Box and Jenkins (1970).  Time Series Analysis.
}

\examples{
data(SeriesB)
plot(SeriesB)
}
\keyword{datasets}
