\name{SeriesB2}
\alias{SeriesB2}
\docType{data}
\title{ IBM Stock Prices, 2nd series}
\description{
Closing price of common stock, daily, June 29 1959 to June 30 1960
}

\usage{data(SeriesB2)}

\format{
  The format is:
 num [1:255] 445 448 450 447 451 453 454 454 459 440 ...
}


\source{
Box and Jenkins (1970).  Time Series Analysis.
}

\examples{
data(SeriesB2)
TimeSeriesPlot(SeriesB2)
}

\keyword{datasets}
